# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CurrentWeekResponse:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      start_date:
        type: string
      week_number:
        type: integer
      winner_count:
        type: integer
    type: object
  dtos.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dtos.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dtos.ThunderSeatResponse:
    properties:
      answer:
        type: string
      created_on:
        type: string
      id:
        type: integer
      media_type:
        type: string
      media_url:
        type: string
      question_id:
        type: integer
      user_id:
        type: string
      week_number:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@thumsup.com
    name: API Support
  description: API documentation for Thums Up Backend Service including Contest Week
    Management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thums Up Backend API
  version: "1.0"
paths:
...
  /thunder-seat:
    post:
      consumes:
      - multipart/form-data
      description: Submit an answer to a Thunder Seat question for the current week
        with optional media file (audio/video). Requires authentication.
      parameters:
      - description: Question ID
        in: formData
        name: question_id
        required: true
        type: integer
      - description: Answer text
        in: formData
        name: answer
        required: true
        type: string
      - description: Optional media file (audio/video, max 100MB)
        in: formData
        name: media_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Answer submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ThunderSeatResponse'
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to submit answer
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Submit Thunder Seat answer
      tags:
      - Thunder Seat
  /thunder-seat/current-week:
    get:
      consumes:
      - application/json
      description: Retrieve the current active contest week details including dates
        and winner count
      produces:
      - application/json
      responses:
        "200":
          description: Current week retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.CurrentWeekResponse'
              type: object
        "404":
          description: No active contest week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get active contest week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get current contest week information
      tags:
      - Thunder Seat
  /thunder-seat/submissions:
    get:
      consumes:
      - application/json
      description: Retrieve all Thunder Seat answer submissions for the authenticated
        user. Requires authentication.
      produces:
      - application/json
      responses:
        "200":
          description: Submissions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.ThunderSeatResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get submissions
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user's Thunder Seat submissions
      tags:
      - Thunder Seat
...
securityDefinitions:
  APIKey:
    description: Admin API key for protected operations
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"